// front/src/modules/control-plane/shared/types.ts
export interface DatasetDiff {
  name?: boolean;
  params?: boolean;
  schema?: boolean;
  type?: boolean;
  public?: boolean;
  managed?: boolean;
}

export interface NamespaceDiff {
  config?: boolean;
  name?: boolean;
}

export interface ProjectDiff {
  config?: boolean;
  name?: boolean;
  description?: boolean;
}

export interface ExperimentDiff {
  config?: boolean;
  name?: boolean;
  description?: boolean;
  dataset_alias?: boolean;
}